requirements.txt
amqp==2.5.2
appdirs==1.4.3
asgiref==3.3.1
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.3.0
autopep8==1.5.4
billiard==3.6.3.0
boto3==1.16.36
botocore==1.19.36
celery==4.4.1
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
coverage==5.0.3
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.3
django-cas-ng==4.1.1
django-cleanup==4.0.0
django-cors-headers==3.2.1
django-cors-middleware==1.5.0
django-environ==0.4.5
django-filter==2.4.0
django-storages==1.10.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
drf-extra-fields==3.0.3
drf-yasg==1.17.1
entrypoints==0.3
et-xmlfile==1.0.1
Faker==0.9.1
filelock==3.0.12
flake8==3.8.4
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.0
importlib-resources==1.5.0
inflection==0.4.0
isort==4.3.21
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
jmespath==0.10.0
kombu==4.6.8
lazy-object-proxy==1.4.3
lml==0.0.9
lxml==4.6.1
MarkupSafe==1.1.1
mccabe==0.6.1
mixer==6.1.3
mock==4.0.2
more-itertools==8.2.0
openapi-codec==1.3.2
openpyxl==3.0.3
packaging==20.1
pdfkit==0.6.1
Pillow==8.0.1
pipenv==2018.11.26
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.8.1
pycodestyle==2.6.0
pyexcel-io==0.5.20
pyexcel-xls==0.5.8
pyexcel-xlsx==0.5.8
pyflakes==2.2.0
PyJWT==1.7.1
pyparsing==2.4.6
pytest==5.3.5
pytest-cov==2.8.1
python-cas==1.5.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
s3transfer==0.3.3
simplejson==3.17.0
six==1.14.0
snowballstemmer==2.0.0
sqlparse==0.3.1
text-unidecode==1.2
toml==0.10.2
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.8
vine==1.3.0
virtualenv==20.1.0
virtualenv-clone==0.5.4
wcwidth==0.1.8
whitenoise==5.0.1
wrapt==1.11.2
xlrd==1.2.0
xlwt==1.3.0
zipp==3.1.0
